refactor the transformer
