+++ packages/conversation-flow/src/transformation/BranchResolver.ts
@@ -4,21 +4,16 @@ import type { IdNode, Message } from '../types';
  * BranchResolver - Handles branch resolution logic
  *
  * Determines which branch should be active based on:
- * 1. metadata.activeBranchId (explicit ID)
- * 2. metadata.activeBranchIndex (explicit index)
- * 3. Inferred from which branch has children
- * 4. Default to first branch
+ * 1. metadata.activeBranchIndex (explicit index)
+ * 2. Inferred from which branch has children
+ * 3. Default to first branch
  */
 export class BranchResolver {
   /**
    * Get active branch ID from IdNode structure (used in contextTree building)
    */
   getActiveBranchId(message: Message, idNode: IdNode): string {
-    // Priority 1: Try to get from metadata.activeBranchId (direct ID)
-    const activeBranchId = (message.metadata as any)?.activeBranchId;
-    if (activeBranchId) return activeBranchId;
-
-    // Priority 2: Try to get from metadata.activeBranchIndex (index-based)
+    // Priority 1: Try to get from metadata.activeBranchIndex (index-based)
     const activeBranchIndex = (message.metadata as any)?.activeBranchIndex;
     if (
       typeof activeBranchIndex === 'number' &&
@@ -28,7 +23,7 @@ export class BranchResolver {
       return idNode.children[activeBranchIndex].id;
     }
 
-    // Priority 3: Infer from which branch has children
+    // Priority 2: Infer from which branch has children
     for (const child of idNode.children) {
       if (child.children.length > 0) {
         return child.id;
@@ -47,13 +42,7 @@ export class BranchResolver {
     childIds: string[],
     childrenMap: Map<string | null, string[]>,
   ): string {
-    // Priority 1: Try to get from metadata.activeBranchId (direct ID)
-    const activeBranchId = (message.metadata as any)?.activeBranchId;
-    if (activeBranchId && childIds.includes(activeBranchId)) {
-      return activeBranchId;
-    }
-
-    // Priority 2: Try to get from metadata.activeBranchIndex (index-based)
+    // Priority 1: Try to get from metadata.activeBranchIndex (index-based)
     const activeBranchIndex = (message.metadata as any)?.activeBranchIndex;
     if (
       typeof activeBranchIndex === 'number' &&
@@ -63,7 +52,7 @@ export class BranchResolver {
       return childIds[activeBranchIndex];
     }
 
-    // Priority 3: Infer from which child has descendants
+    // Priority 2: Infer from which child has descendants
     for (const childId of childIds) {
       const descendants = childrenMap.get(childId);
       if (descendants && descendants.length > 0) {
