+++ packages/conversation-flow/src/transformation/FlatListBuilder.ts
@@ -1,0 +1,403 @@
+import type { AssistantContentBlock, ChatToolPayloadWithResult } from '@lobechat/types';
+
+import type { Message, MessageGroupMetadata } from '../types';
+import type { BranchResolver } from './BranchResolver';
+import type { MessageCollector } from './MessageCollector';
+import type { MessageTransformer } from './MessageTransformer';
+
+/**
+ * FlatListBuilder - Builds flat message list following the active path
+ *
+ * Handles:
+ * 1. Recursive traversal following active branches
+ * 2. Creating virtual messages for Compare and AssistantGroup
+ * 3. Processing different message types with priority
+ */
+export class FlatListBuilder {
+  constructor(
+    private messageMap: Map<string, Message>,
+    private messageGroupMap: Map<string, MessageGroupMetadata>,
+    private childrenMap: Map<string | null, string[]>,
+    private branchResolver: BranchResolver,
+    private messageCollector: MessageCollector,
+    private messageTransformer: MessageTransformer,
+  ) {}
+
+  /**
+   * Generate flatList from messages array
+   * Only includes messages in the active path
+   */
+  flatten(messages: Message[]): Message[] {
+    const flatList: Message[] = [];
+    const processedIds = new Set<string>();
+
+    // Build the active path by traversing from root
+    this.buildFlatListRecursive(null, flatList, processedIds, messages);
+
+    return flatList;
+  }
+
+  /**
+   * Recursively build flatList following the active path
+   */
+  private buildFlatListRecursive(
+    parentId: string | null,
+    flatList: Message[],
+    processedIds: Set<string>,
+    allMessages: Message[],
+  ): void {
+    const children = this.childrenMap.get(parentId) ?? [];
+
+    for (const childId of children) {
+      if (processedIds.has(childId)) continue;
+
+      const message = this.messageMap.get(childId);
+      if (!message) continue;
+
+      // Priority 1: Compare message group
+      const messageGroup = message.groupId ? this.messageGroupMap.get(message.groupId) : undefined;
+
+      if (messageGroup && messageGroup.mode === 'compare' && !processedIds.has(messageGroup.id)) {
+        const groupMembers = this.messageCollector.collectGroupMembers(
+          message.groupId!,
+          allMessages,
+        );
+        const compareMessage = this.createCompareMessage(messageGroup, groupMembers);
+        flatList.push(compareMessage);
+        groupMembers.forEach((m) => processedIds.add(m.id));
+        processedIds.add(messageGroup.id);
+
+        // Continue with active column's children (if any)
+        if ((compareMessage as any).activeColumnId) {
+          this.buildFlatListRecursive(
+            (compareMessage as any).activeColumnId,
+            flatList,
+            processedIds,
+            allMessages,
+          );
+        }
+        continue;
+      }
+
+      // Priority 2: AssistantGroup (assistant + tools)
+      if (message.role === 'assistant' && message.tools && message.tools.length > 0) {
+        // Collect the entire assistant group chain
+        const assistantChain: Message[] = [];
+        const allToolMessages: Message[] = [];
+        this.messageCollector.collectAssistantChain(
+          message,
+          allMessages,
+          assistantChain,
+          allToolMessages,
+          processedIds,
+        );
+
+        // Create assistantGroup virtual message
+        const groupMessage = this.createAssistantGroupMessage(
+          assistantChain[0],
+          assistantChain,
+          allToolMessages,
+        );
+        flatList.push(groupMessage);
+
+        // Mark all as processed
+        assistantChain.forEach((m) => processedIds.add(m.id));
+        allToolMessages.forEach((m) => processedIds.add(m.id));
+
+        // Continue after the last tool message
+        const lastToolMsg = allToolMessages.at(-1);
+        if (lastToolMsg) {
+          this.buildFlatListRecursive(lastToolMsg.id, flatList, processedIds, allMessages);
+        }
+        continue;
+      }
+
+      // Priority 3a: Compare mode from user message metadata
+      const childMessages = this.childrenMap.get(message.id) ?? [];
+      if (this.isCompareMode(message) && childMessages.length > 1) {
+        // Add user message
+        flatList.push(message);
+        processedIds.add(message.id);
+
+        // Create compare virtual message with proper handling of AssistantGroups
+        const compareMessage = this.createCompareMessageFromChildIds(
+          message,
+          childMessages,
+          allMessages,
+          processedIds,
+        );
+        flatList.push(compareMessage);
+
+        // Continue with active column's children (if any)
+        if ((compareMessage as any).activeColumnId) {
+          this.buildFlatListRecursive(
+            (compareMessage as any).activeColumnId,
+            flatList,
+            processedIds,
+            allMessages,
+          );
+        }
+        continue;
+      }
+
+      // Priority 3b: User message with branches
+      if (message.role === 'user' && childMessages.length > 1) {
+        const activeBranchId = this.branchResolver.getActiveBranchIdFromMetadata(
+          message,
+          childMessages,
+          this.childrenMap,
+        );
+        const userWithBranches = this.createUserMessageWithBranches(message);
+        flatList.push(userWithBranches);
+        processedIds.add(message.id);
+
+        // Continue with active branch and process its message
+        const activeBranchMsg = this.messageMap.get(activeBranchId);
+        if (activeBranchMsg) {
+          flatList.push(activeBranchMsg);
+          processedIds.add(activeBranchId);
+
+          // Continue with active branch's children
+          this.buildFlatListRecursive(activeBranchId, flatList, processedIds, allMessages);
+        }
+        continue;
+      }
+
+      // Priority 3c: Assistant message with branches
+      if (message.role === 'assistant' && childMessages.length > 1) {
+        const activeBranchId = this.branchResolver.getActiveBranchIdFromMetadata(
+          message,
+          childMessages,
+          this.childrenMap,
+        );
+        // Add the assistant message itself
+        flatList.push(message);
+        processedIds.add(message.id);
+
+        // Continue with active branch and process its message
+        const activeBranchMsg = this.messageMap.get(activeBranchId);
+        if (activeBranchMsg) {
+          flatList.push(activeBranchMsg);
+          processedIds.add(activeBranchId);
+
+          // Continue with active branch's children
+          this.buildFlatListRecursive(activeBranchId, flatList, processedIds, allMessages);
+        }
+        continue;
+      }
+
+      // Priority 4: Regular message
+      flatList.push(message);
+      processedIds.add(message.id);
+
+      // Continue with children
+      this.buildFlatListRecursive(message.id, flatList, processedIds, allMessages);
+    }
+  }
+
+  /**
+   * Check if message has compare mode in metadata
+   */
+  private isCompareMode(message: Message): boolean {
+    return (message.metadata as any)?.compare === true;
+  }
+
+  /**
+   * Create compare virtual message from child IDs with AssistantGroup support
+   */
+  private createCompareMessageFromChildIds(
+    parentMessage: Message,
+    childIds: string[],
+    allMessages: Message[],
+    processedIds: Set<string>,
+  ): Message {
+    const columns: Message[][] = [];
+    const columnFirstIds: string[] = [];
+    let activeColumnId: string | undefined;
+
+    // Process each child (column)
+    for (const childId of childIds) {
+      const childMessage = this.messageMap.get(childId);
+      if (!childMessage) continue;
+
+      columnFirstIds.push(childId);
+
+      // Check if this message is marked as active column
+      if ((childMessage.metadata as any)?.activeColumn === true) {
+        activeColumnId = childId;
+      }
+
+      // Check if this child is an AssistantGroup
+      if (
+        childMessage.role === 'assistant' &&
+        childMessage.tools &&
+        childMessage.tools.length > 0
+      ) {
+        // Collect the entire assistant group chain for this column
+        const assistantChain: Message[] = [];
+        const allToolMessages: Message[] = [];
+        const columnProcessedIds = new Set<string>();
+
+        this.messageCollector.collectAssistantChain(
+          childMessage,
+          allMessages,
+          assistantChain,
+          allToolMessages,
+          columnProcessedIds,
+        );
+
+        // Create assistantGroup virtual message for this column
+        const groupMessage = this.createAssistantGroupMessage(
+          assistantChain[0],
+          assistantChain,
+          allToolMessages,
+        );
+
+        columns.push([groupMessage]);
+
+        // Mark all as processed
+        assistantChain.forEach((m) => processedIds.add(m.id));
+        allToolMessages.forEach((m) => processedIds.add(m.id));
+      } else {
+        // Regular message (not an AssistantGroup)
+        columns.push([childMessage]);
+        processedIds.add(childId);
+      }
+    }
+
+    // Generate ID with all column first message IDs
+    const columnIdsStr = columnFirstIds.join('-');
+    const compareId = `compare-${parentMessage.id}-${columnIdsStr}`;
+
+    // Calculate timestamps from first column's messages
+    const firstColumnMessages = childIds.map((id) => this.messageMap.get(id)).filter(Boolean);
+    const createdAt =
+      firstColumnMessages.length > 0
+        ? Math.min(...firstColumnMessages.map((m) => m!.createdAt))
+        : parentMessage.createdAt;
+    const updatedAt =
+      firstColumnMessages.length > 0
+        ? Math.max(...firstColumnMessages.map((m) => m!.updatedAt))
+        : parentMessage.updatedAt;
+
+    return {
+      activeColumnId,
+      columns: columns as any,
+      content: '',
+      createdAt,
+      extra: {
+        parentMessageId: parentMessage.id,
+      },
+      id: compareId,
+      meta: parentMessage.meta || {},
+      role: 'compare' as any,
+      updatedAt,
+    } as Message;
+  }
+
+  /**
+   * Create compare virtual message (for group-based compare)
+   */
+  private createCompareMessage(group: MessageGroupMetadata, members: Message[]): Message {
+    // Find active column ID from members metadata
+    const activeColumnId = members.find((msg) => (msg.metadata as any)?.activeColumn === true)?.id;
+
+    // columns contain full Message objects
+    const columns: Message[][] = members.map((msg) => [msg]);
+
+    return {
+      activeColumnId,
+      columns: columns as any,
+      content: '',
+      createdAt: Math.min(...members.map((m) => m.createdAt)),
+      extra: {
+        groupMode: group.mode,
+        parentMessageId: group.parentMessageId,
+      },
+      id: group.id,
+      meta: members[0]?.meta || {},
+      role: 'compare' as any,
+      updatedAt: Math.max(...members.map((m) => m.updatedAt)),
+    } as Message;
+  }
+
+  /**
+   * Create assistant group virtual message from entire chain
+   */
+  private createAssistantGroupMessage(
+    firstAssistant: Message,
+    assistantChain: Message[],
+    allToolMessages: Message[],
+  ): Message {
+    const children: AssistantContentBlock[] = [];
+
+    // Create tool map for lookup
+    const toolMap = new Map<string, Message>();
+    allToolMessages.forEach((tm) => {
+      if (tm.tool_call_id) {
+        toolMap.set(tm.tool_call_id, tm);
+      }
+    });
+
+    // Process each assistant in the chain
+    for (const assistant of assistantChain) {
+      // Build toolsWithResults for this assistant
+      const toolsWithResults: ChatToolPayloadWithResult[] =
+        assistant.tools?.map((tool) => {
+          const toolMsg = toolMap.get(tool.id);
+          if (toolMsg) {
+            return {
+              ...tool,
+              result: {
+                content: toolMsg.content || '',
+                error: toolMsg.error,
+                id: toolMsg.id,
+                state: toolMsg.pluginState,
+              },
+              result_msg_id: toolMsg.id,
+            };
+          }
+          return tool;
+        }) || [];
+
+      const { usage: msgUsage, performance: msgPerformance } =
+        this.messageTransformer.splitMetadata(assistant.metadata);
+
+      children.push({
+        content: assistant.content || '',
+        error: assistant.error,
+        id: assistant.id,
+        imageList:
+          assistant.imageList && assistant.imageList.length > 0 ? assistant.imageList : undefined,
+        performance: msgPerformance,
+        reasoning: assistant.reasoning || undefined,
+        tools: toolsWithResults.length > 0 ? toolsWithResults : undefined,
+        usage: msgUsage,
+      });
+    }
+
+    const aggregated = this.messageTransformer.aggregateMetadata(children);
+
+    return {
+      ...firstAssistant,
+      children,
+      content: '',
+      imageList: undefined,
+      metadata: undefined,
+      performance: aggregated.performance,
+      reasoning: undefined,
+      role: 'assistantGroup' as any,
+      tools: undefined,
+      usage: aggregated.usage,
+    };
+  }
+
+  /**
+   * Create user message with branch metadata
+   */
+  private createUserMessageWithBranches(user: Message): Message {
+    // Just return the original user message with its metadata.activeBranchId
+    // No need to add extra.branches
+    return { ...user };
+  }
+}
