+++ packages/conversation-flow/src/transformation/__tests__/BranchResolver.test.ts
@@ -1,35 +1,12 @@
 import { describe, expect, it } from 'vitest';
 
-import { BranchResolver } from '../../transformation/BranchResolver';
 import type { IdNode, Message } from '../../types';
+import { BranchResolver } from '../BranchResolver';
 
 describe('BranchResolver', () => {
   const resolver = new BranchResolver();
 
   describe('getActiveBranchId', () => {
-    it('should return activeBranchId from metadata when present', () => {
-      const message: Message = {
-        content: 'test',
-        createdAt: 0,
-        id: 'msg-1',
-        meta: {},
-        metadata: { activeBranchId: 'msg-3' },
-        parentId: null,
-        role: 'user',
-        updatedAt: 0,
-      };
-
-      const idNode: IdNode = {
-        children: [
-          { children: [], id: 'msg-2' },
-          { children: [], id: 'msg-3' },
-        ],
-        id: 'msg-1',
-      };
-
-      expect(resolver.getActiveBranchId(message, idNode)).toBe('msg-3');
-    });
-
     it('should return branch by activeBranchIndex from metadata when present', () => {
       const message: Message = {
         content: 'test',
@@ -37,7 +14,6 @@ describe('BranchResolver', () => {
         id: 'msg-1',
         meta: {},
         metadata: { activeBranchIndex: 1 },
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -59,7 +35,6 @@ describe('BranchResolver', () => {
         createdAt: 0,
         id: 'msg-1',
         meta: {},
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -81,7 +56,6 @@ describe('BranchResolver', () => {
         createdAt: 0,
         id: 'msg-1',
         meta: {},
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -104,7 +78,6 @@ describe('BranchResolver', () => {
         id: 'msg-1',
         meta: {},
         metadata: { activeBranchIndex: 5 }, // out of bounds
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -123,24 +96,6 @@ describe('BranchResolver', () => {
   });
 
   describe('getActiveBranchIdFromMetadata', () => {
-    it('should return activeBranchId from metadata when present and valid', () => {
-      const message: Message = {
-        content: 'test',
-        createdAt: 0,
-        id: 'msg-1',
-        meta: {},
-        metadata: { activeBranchId: 'msg-3' },
-        parentId: null,
-        role: 'user',
-        updatedAt: 0,
-      };
-
-      const childIds = ['msg-2', 'msg-3'];
-      const childrenMap = new Map<string | null, string[]>();
-
-      expect(resolver.getActiveBranchIdFromMetadata(message, childIds, childrenMap)).toBe('msg-3');
-    });
-
     it('should return branch by activeBranchIndex from metadata', () => {
       const message: Message = {
         content: 'test',
@@ -148,7 +103,6 @@ describe('BranchResolver', () => {
         id: 'msg-1',
         meta: {},
         metadata: { activeBranchIndex: 1 },
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -165,7 +119,6 @@ describe('BranchResolver', () => {
         createdAt: 0,
         id: 'msg-1',
         meta: {},
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -185,7 +138,6 @@ describe('BranchResolver', () => {
         createdAt: 0,
         id: 'msg-1',
         meta: {},
-        parentId: null,
         role: 'user',
         updatedAt: 0,
       };
@@ -193,25 +145,6 @@ describe('BranchResolver', () => {
       const childIds = ['msg-2', 'msg-3'];
       const childrenMap = new Map<string | null, string[]>();
 
-      expect(resolver.getActiveBranchIdFromMetadata(message, childIds, childrenMap)).toBe('msg-2');
-    });
-
-    it('should ignore activeBranchId if not in childIds', () => {
-      const message: Message = {
-        content: 'test',
-        createdAt: 0,
-        id: 'msg-1',
-        meta: {},
-        metadata: { activeBranchId: 'msg-999' }, // not in childIds
-        parentId: null,
-        role: 'user',
-        updatedAt: 0,
-      };
-
-      const childIds = ['msg-2', 'msg-3'];
-      const childrenMap = new Map<string | null, string[]>();
-
-      // Should default to first child
       expect(resolver.getActiveBranchIdFromMetadata(message, childIds, childrenMap)).toBe('msg-2');
     });
   });
