+++ src/services/message/index.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import {
   ChatMessageError,
   ChatMessagePluginError,
